feat(core): add client utilities file with one utility function
The client utilities file in the core package is the entry point to
the shared client side logic of the inference client which is pure
JavaScript code that can be used across all framework-specific
packages.

Create and export  inside this file one function for now which is
`promptRequest` that handles taking and sending the prompts
of the users to the server side for processing, reading the streaming
response of the LLMs and providing it token by token via `onData`
callback funtion.