feat(core): create the server side logic of the inference client
Create `src/lib/server.ts` file which is the entry point to the server
side logic of the Mecha Agent inference client. And inside this file
create four core functionalities:

1. `GetAgentData` function, Which fetches the data of the published
agent that is supposed to be accessed by the public.

2. `SendPrompt` function, Which is responsible for sending the prompts
of the users and getting responses from mecha agent server safely.

3. `GetChatMessages` function, Which fetches the messages of the
last chat session (if there is one) to continue the chat from where the
user stopped.

4. `GetAgentAvatar` function, Which is responsible forr getting the avatar
of the agent.

Therse functions are supposed to be run in a server side environment for
securety reasons like keeping the API Key that authorizes its holder to access
things he is not allowed to access in a secure environment.