feat(nextjs): create the entry point component to the client interface
Create `MechaAgentChat` component which is the entry point to the
mecha agent client interface and the only component that should be
exposed to the end users to put it in their web apps.

Create `/src/main.ts` file and export inside it the routes handler that
handles the backend part of this package and `MechaAgentChat`
component that represents the frontend part.

Add `exports` field to the `package.json` file of this package and
specify from where end users can import the content of the package.