feat(sveltekit): create the entry point component to the client interface
Create `MechaAgentChat` component which is the entry point to the
mecha agent client interface and the only UI component that should
be exposed to the end users to put it in their web apps.

Create `/src/lib/index.ts` file and export inside it the `MechaAgentChat`
component that represents the frontend part of this package.

Add `exports`, `types` and `svelte` fields to the `package.json` file
of this package to specify how consumers can consume the content
of the package.